+++ ZeldaLoot_Classic.lua
@@ -1,4 +1,4 @@
-AddonVersion = "|cff00ff001.2.5|r"
+AddonVersion = "|cff00ff001.3.0|r"
 ZL_AddonName = "ZeldaLoot Classic"
 ZL_AddonColor = "|cff00ffff"
 ZL_soundHandle = 0
@@ -10,8 +10,10 @@ end
 function play_zeldaSound(index, sound_file)
 	local sound_set = get_sound_set(index)
 	local willPlay = nil
-	local sound_ext = "ogg"
-	local sound_channel = "SFX"
+	local sound_ext = get_sound_type()
+	local sound_channel = get_sound_channel()
+	local warning_text = ""
+
 	if (ZL_soundHandle ~= 0) then
 		if (zl_debug_bool) then
 			zl_Print("Stopping sound for " .. ZL_soundHandle)
@@ -19,16 +21,18 @@ function play_zeldaSound(index, sound_file)
 		StopSound(ZL_soundHandle, 0)
 	end
 
+	if (zl_warning_bool) then
+		warning_text = "WARNING "
+	end
 
 	update_config(false)
 	willPlay, ZL_soundHandle = PlaySoundFile("Interface\\AddOns\\ZeldaLoot_Classic\\Sounds\\Sets\\" .. sound_set .. "\\" .. sound_file .. "." .. sound_ext, sound_channel)
-	if (zl_debug_bool) then
-		local mess = "[" .. sound_set .. "\\" .. sound_file .. "." .. sound_ext .. "] on sound channel [" .. sound_channel .. "]"
-		if (willPlay) then
-			zl_Print("Playing sound for " .. mess)
-		else
-			zl_Print("NOT playing sound for " .. mess .. " likey due to [" .. sound_channel .. "] being muted")
-		end
+
+	local mess = "[" .. sound_set .. "\\" .. sound_file .. "." .. sound_ext .. "] on sound channel [" .. sound_channel .. "]"
+	if (willPlay and zl_debug_bool) then
+		zl_Print("Playing sound for " .. mess)
+	elseif (zl_warning_bool or zl_debug_bool) then
+		zl_Print(warning_text .. "NOT playing sound for " .. mess .. " likey due to [" .. sound_channel .. "] being muted")
 	end
 end
 
@@ -153,6 +157,11 @@ function reset_config(print_text)
 
 		inherited = {
 			include  = true
+		},
+
+		settings = {
+			ext = "wav",
+			channel = "SFX"
 		}
 	}